CapsuleForensics/requirements.txt
numpy
torch-1.4.0-cp37-cp37m-manylinux1_x86_64.whl
tqdm
scikit-learn
